% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/allocateDataToNodes.R
\name{allocateDataToNodes}
\alias{allocateDataToNodes}
\title{Allocate data to Terminal nodes}
\usage{
allocateDataToNodes(i_data, i_nodeInfo, DEBUG = FALSE)
}
\description{
This function allocates data to terminal nodes
}
\keyword{allocation,}
\keyword{data,}
\keyword{node,}
\keyword{regression}
\keyword{terminal}
\keyword{tree}
