% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/as.fake.rpart.tree.R
\name{as.fake.rpart.tree}
\alias{as.fake.rpart.tree}
\title{Creates fake rpart tree object}
\usage{
as.fake.rpart.tree(i_gtree)
}
\description{
This function creates a fake rpart tree object for plotting purposes
}
\keyword{GUIDE,}
\keyword{plot,}
\keyword{prp}
\keyword{tree,}
