% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/calc_transitionMatrix.R
\name{calc_transitionMatrix}
\alias{calc_transitionMatrix}
\title{Compute transition matrix}
\usage{
calc_transitionMatrix(i_alloc, i_nodeInfo)
}
\description{
This function computes a transition matrix between terminal nodes
}
\keyword{data,}
\keyword{matrix,}
\keyword{node,}
\keyword{regression}
\keyword{terminal}
\keyword{transition}
\keyword{tree}
