% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/catret.R
\name{catret}
\alias{catret}
\title{Print message with return carriage}
\usage{
catret(i_text, ...)
}
\description{
This function is a modification of the original cat function that appends a return carriage character at the end of a submitted string
}
\keyword{carriage}
\keyword{cat,}
\keyword{message,}
\keyword{return}
