% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/compare_treeStructure.R
\name{compare_treeStructure}
\alias{compare_treeStructure}
\title{Compare tree structures}
\usage{
compare_treeStructure(i_tree1, i_tree2, ...)
}
\description{
This function computes a similarity index based on the edit distance of 2 tree structures
}
\keyword{index}
\keyword{regression}
\keyword{similarity}
\keyword{tree,}
