% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/compute_editDistance.R
\name{compute_editDistance}
\alias{compute_editDistance}
\title{Compute edit distance}
\usage{
compute_editDistance(v1, v2, SORT = TRUE, NORMALISE = FALSE)
}
\description{
This function computes the edit distance of two vectors of items
}
\keyword{distance}
\keyword{edit}
\keyword{regression}
\keyword{tree,}
