% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/generate_input.R
\name{generate_input}
\alias{generate_input}
\title{Generate GUIDE input data}
\usage{
generate_input(i_data, i_vardef, i_treeType, i_treeOptions, i_fpath,
  i_missVal = "NaN", i_startLine = "2", i_idcolumns = c("iso", "year"),
  VERBOSE = FALSE)
}
\description{
This function generates the data and other files required to run GUIDE
}
\keyword{GUIDE,}
\keyword{data}
\keyword{description}
\keyword{file}
\keyword{file,}
\keyword{input,}
