% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_pode.R
\name{get_pode}
\alias{get_pode}
\title{Get the proportion of deviance explained}
\usage{
get_pode(outfpath)
}
\description{
This function get the proportion of deviance explained from a GUIDE output file.
}
\keyword{GUIDE,}
\keyword{output,}
\keyword{parse,}
\keyword{pode}
