% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_secondBestSplit.R
\name{get_secondBestSplit}
\alias{get_secondBestSplit}
\title{Get the second best split variable}
\usage{
get_secondBestSplit(outfpath, REMOVE_LOG = FALSE)
}
\description{
This function get the second best split variable from a GUIDE output file.
}
\keyword{GUIDE,}
\keyword{best}
\keyword{output,}
\keyword{parse,}
\keyword{second}
\keyword{split}
