% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_warnings.R
\name{get_warnings}
\alias{get_warnings}
\title{Get warnings from GUIDE output}
\usage{
get_warnings(outfpath)
}
\description{
This function extract warning messages in a GUIDE output file.
}
\keyword{GUIDE,}
\keyword{output,}
\keyword{parse,}
\keyword{warning}
