% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/getCountryPathway.R
\name{getCountryPathway}
\alias{getCountryPathway}
\title{Get country pathway of ISO}
\usage{
getCountryPathway(i_alloc, i_iso)
}
\description{
This function generates a scatter plot
}
\keyword{country,}
\keyword{pathway,}
\keyword{regression}
\keyword{tree}
