% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/getParentNodeID.R
\name{getParentNodeID}
\alias{getParentNodeID}
\title{Get complete tree branch paths for a given terminal node}
\usage{
getParentNodeID(i_nodeID, i_nodeInfo)
}
\description{
This recursive function to get the intermediate nodes across Root-Terminal paths (i.e. From a Root node to a Terminal node)
}
\keyword{branch,}
\keyword{path}
\keyword{regression}
\keyword{tree,}
