% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/getPathNodeInfo.R
\name{getPathNodeInfo}
\alias{getPathNodeInfo}
\title{Get full path of from root node to terminal node}
\usage{
getPathNodeInfo(i_tnode, i_nodeInfo, DEBUG = FALSE)
}
\description{
This function generates a composite plot for a given terminal node
}
\keyword{node,}
\keyword{plot,}
\keyword{regression}
\keyword{terminal}
\keyword{tree}
