% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/getSplitVarPath.R
\name{getSplitVarPath}
\alias{getSplitVarPath}
\title{Get tree branch paths of specific depth (from root node)}
\usage{
getSplitVarPath(i_nodeInfo, i_depth, DIGITS = NULL, REMOVE_LOG = FALSE,
  REMOVE_MEAN = FALSE)
}
\description{
This function
}
\keyword{branch,}
\keyword{path}
\keyword{regression}
\keyword{tree,}
