% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/initpath.R
\name{initpath}
\alias{initpath}
\title{Initialise file paths}
\usage{
initpath(outpath, commonName = "GUIDEfile")
}
\description{
This function initialises file paths.
}
\keyword{GUIDE,}
\keyword{directory,}
\keyword{file,}
\keyword{folder,}
\keyword{path}
