% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/initTreeOptions.R
\name{initTreeOptions}
\alias{initTreeOptions}
\title{Initialise tree options}
\usage{
initTreeOptions(prune = "1", nbcv = 10, cvtype = "1", sevalue = 0.5,
  search = "2", quantile1 = 0.25, quantile2 = 0.75, splitfrac = 1,
  truncmeth = "3", missregval = "2", maxsplits = 10, minnbnodes = 3)
}
\description{
This function initialises tree options.
}
\keyword{GUIDE,}
\keyword{model,}
\keyword{options}
\keyword{tree,}
