% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/multiplot.R
\name{multiplot}
\alias{multiplot}
\title{Multiple plot function}
\usage{
multiplot(..., plotlist = NULL, file, cols = 1, layout = NULL)
}
\description{
This function generates a composite plot for a given terminal node
}
\keyword{node,}
\keyword{plot,}
\keyword{regression}
\keyword{terminal}
\keyword{tree}
