% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/parse_output.R
\name{parse_output}
\alias{parse_output}
\title{Parse GUIDE output}
\usage{
parse_output(outfpath, vardef)
}
\description{
This function get tree structure information from a GUIDE output file
}
\keyword{GUIDE,}
\keyword{output,}
\keyword{parse}
