% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/parse_regcoef.R
\name{parse_regcoef}
\alias{parse_regcoef}
\title{Parse GUIDE regression coefficient file}
\usage{
parse_regcoef(outfpath)
}
\description{
This function read the GUIDE output file containing regression coefficients
}
\keyword{GUIDE,}
\keyword{coefficients,}
\keyword{output,}
\keyword{parse}
\keyword{regression}
