% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/parse_regressor.R
\name{parse_regressor}
\alias{parse_regressor}
\title{Parse GUIDE regressors}
\usage{
parse_regressor(outfpath)
}
\description{
This function parses a GUIDE regressor file.
}
\keyword{GUIDE,}
\keyword{output,}
\keyword{parse}
\keyword{regressor,}
