% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/perform_regression_diagnostics.R
\name{perform_regression_diagnostics}
\alias{perform_regression_diagnostics}
\title{Perform regression diagnostics}
\usage{
perform_regression_diagnostics(i_data, i_dvar, i_idvar)
}
\description{
This function performs a set of diagnostics to assess the performance of a linear regression model
}
\keyword{diagnostics}
\keyword{regression,}
