% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_ISOPathway.R
\name{plot_ISOPathway}
\alias{plot_ISOPathway}
\title{Plot pathwas of a single ISO}
\usage{
plot_ISOPathway(i_iso, i_alloc, i_nodeInfo)
}
\description{
This function generates a scatter plot
}
\keyword{pathway,}
\keyword{plot,}
\keyword{regression}
\keyword{tree}
