% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_ISOpathways.R
\name{plot_ISOPathways}
\alias{plot_ISOPathways}
\title{Plot pathwas of one or several ISOs}
\usage{
plot_ISOPathways(i_iso, i_alloc, i_tree, export = NULL)
}
\description{
This function generates a network plot
}
\keyword{pathway,}
\keyword{plot,}
\keyword{regression}
\keyword{tree}
