% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_ScatterCountry.R
\name{plot_ScatterCountry}
\alias{plot_ScatterCountry}
\title{Generate a scatter plot for a particular country}
\usage{
plot_ScatterCountry(i_alloc, i_nodeInfo, i_iso, i_var_x, i_var_y, i_var_col,
  PLOT_LABELS = FALSE)
}
\description{
This function generates a scatter plot
}
\keyword{plot,}
\keyword{regression}
\keyword{scatter,}
\keyword{tree}
