% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_ScatterCountry_allVars.R
\name{plot_ScatterCountry_allVars}
\alias{plot_ScatterCountry_allVars}
\title{Generate scatter pots for a particular country}
\usage{
plot_ScatterCountry_allVars(i_alloc, i_nodeInfo, i_iso, i_var_dep,
  PLOT_LABELS = FALSE)
}
\description{
This function generates a scatter plot for all variables
}
\keyword{plot,}
\keyword{regression}
\keyword{scatter,}
\keyword{tree}
