% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_Tnode.R
\name{plot_TNode}
\alias{plot_TNode}
\title{Plot terminal node}
\usage{
plot_TNode(i_alloc, i_node, i_tree, PLOT_MAP = FALSE,
  PLOT_TRANSITION = FALSE)
}
\description{
This function generates a composite plot for a given terminal node
}
\keyword{node,}
\keyword{plot,}
\keyword{regression}
\keyword{terminal}
\keyword{tree}
