% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/run_guide.R
\name{run_guide}
\alias{run_guide}
\title{run GUIDE}
\usage{
run_guide(i_fpath, IMPORTANCE_SCORING = FALSE, SHOW_LOG = FALSE,
  DEBUG = FALSE)
}
\description{
This function calls GUIDE with a provided batch input file.
}
\keyword{GUIDE,}
\keyword{run}
