% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/write_batch_file.R
\name{write_batch_file}
\alias{write_batch_file}
\title{Create a GUIDE batch input file}
\usage{
write_batch_file(i_singleTree, i_treeType, i_treeOptions, i_fpath,
  VERBOSE = FALSE)
}
\description{
This function generates a GUIDE batch input file
}
\keyword{GUIDE,}
\keyword{batch}
\keyword{file}
\keyword{input,}
