% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/write_is_file.R
\name{write_is_file}
\alias{write_is_file}
\title{Create a GUIDE importance scoring batch input file}
\usage{
write_is_file(i_singleTree, i_treeType, i_treeOptions, i_fpath,
  VERBOSE = FALSE)
}
\description{
This function generates a GUIDE importance scoring batch input file
}
\keyword{GUIDE,}
\keyword{batch}
\keyword{file}
\keyword{input,}
